// front-end/src/app/models/payment/payment-request.model.ts
import { AttachedDocumentModel } from "./attached-document.model";

export class PaymentRequestModel {
    idSolicitacao: number;
    statusGestor: number | null;
    statusContabilidade: number | null;
    statusFinanceiro: number | null;
    idUsuario: number;
    nomeUsuario: string | null;
    idDepartamento: number;
    idUsuarioAtribuido: number | null;
    dataAtribuicao: Date | null;
    idUsuarioAtribuidor: number | null;
    dataSolicitacao: Date;
    tipoSolicitacao: number;
    tipoPagamento: number;
    dataDocumento: Date | null;
    dataVencimento: Date | null;
    numeroDocumento: string | null;
    descricao: string;
    valor: number;
    parcelado: boolean | null;
    parcelas: number | null;
    formaPagamento: number;
    centroCusto: string;
    projeto: string;
    pca: string;
    numDocParceiro: string;
    codigoParceiro: string;
    nomeParceiro: string;
    bancoParceiro: string;
    agenciaParceiro: string;
    contaParceiro: string;
    observacao: string;
    tipoAutorizacao: number;
    idGestor: number | null;
    nomeGestor: string | null;
    aprovadoGestor: boolean | null;
    observacaoGestor: string;
    aprovadoSetor: boolean | null;
    observacaoSetor: string;
    documentos: AttachedDocumentModel[] | null;
    gravarDadosParceiro: boolean | null;
    dataAprovacaoGestor: string | null;
    dataAprovacaoFinanceiro: string | null;
    dataAprovacaoContabil: string | null;
    rascunho: boolean = true;
    cancelado: boolean = false;
}